Move get_stat_data core implementation
