--- wazimap/data/__init__.py
@@ -1,4 +1,0 @@
-from .utils import LocationNotFound
-
-
-__all__ = ['LocationNotFound']
